move the promo dialog to the My Sites screen
